requirements.txt
fastapi
uvicorn
python-multipart
google-genai
Pillow
mysql-connector-python
sqlalchemy
passlib[bcrypt]
python-jose[cryptography]
python-dotenv